02_Aplicacion/Backend/Digiturno_APE_Backend_Python/requirements.txt
uvicorn==0.30.6
fastapi==0.115.6
selenium==4.27.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.28.2
webdriver-manager==4.0.2

